mds: move MDSCacheObject related code to seperate files
