+++ src/mds/CMakeLists.txt
@@ -33,6 +33,7 @@ set(mds_srcs
   MDSContext.cc
   MDSAuthCaps.cc
   MDLog.cc
+  MDSCacheObject.cc
   ${CMAKE_SOURCE_DIR}/src/common/TrackedOp.cc
   ${CMAKE_SOURCE_DIR}/src/osdc/Journaler.cc)
 add_library(mds STATIC ${mds_srcs}
